+++ packages/mcp/package.json
@@ -47,6 +47,7 @@
         "typecheck": "tsc -b --emitDeclarationOnly"
     },
     "dependencies": {
+        "@gtkx/utils": "workspace:*",
         "@modelcontextprotocol/sdk": "^1.29.0",
         "zod": "^4.4.3"
     }
