+++ packages/mcp/src/index.ts
@@ -1,16 +1,4 @@
-export {
-    exitCodeForSignal,
-    type GracefulShutdownHandle,
-    type GracefulShutdownOptions,
-    installGracefulShutdown,
-} from "./graceful-shutdown.js";
-export {
-    errorMessage,
-    McpError,
-    McpErrorCode,
-    methodNotFoundError,
-    widgetNotFoundError,
-} from "./protocol/errors.js";
+export { McpError, McpErrorCode, methodNotFoundError, widgetNotFoundError } from "./protocol/errors.js";
 export {
     DEFAULT_SOCKET_PATH,
     type IpcMethod,
