+++ packages/utils/package.json
@@ -1,18 +1,8 @@
 {
-    "name": "@gtkx/vitest",
-    "version": "0.21.0",
-    "description": "Vitest plugin for GTKX applications with Xvfb display isolation",
-    "keywords": [
-        "gtkx",
-        "gtk",
-        "gtk4",
-        "vitest",
-        "testing",
-        "plugin",
-        "xvfb",
-        "linux",
-        "desktop"
-    ],
+    "name": "@gtkx/utils",
+    "version": "0.0.0",
+    "private": true,
+    "description": "Cross-package utilities shared by GTKX runtime entry points",
     "homepage": "https://gtkx.dev",
     "bugs": {
         "url": "https://github.com/gtkx-org/gtkx/issues"
@@ -20,7 +10,7 @@
     "repository": {
         "type": "git",
         "url": "https://github.com/gtkx-org/gtkx.git",
-        "directory": "packages/vitest"
+        "directory": "packages/utils"
     },
     "license": "MPL-2.0",
     "author": "Eugenio Depalo <eugeniodepalo@gmail.com>",
@@ -28,6 +18,7 @@
     "exports": {
         "./package.json": "./package.json",
         ".": {
+            "source": "./src/index.ts",
             "types": "./dist/index.d.ts",
             "default": "./dist/index.js"
         }
@@ -39,15 +30,7 @@
     ],
     "scripts": {
         "build": "tsc -b",
+        "test": "vitest run",
         "typecheck": "tsc -b --emitDeclarationOnly"
-    },
-    "dependencies": {
-        "@gtkx/utils": "workspace:*"
-    },
-    "devDependencies": {
-        "vitest": "^4.1.7"
-    },
-    "peerDependencies": {
-        "vitest": ">=4"
     }
 }
