+++ packages/utils/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "files": [],
+    "include": [],
+    "references": [
+        {
+            "path": "tsconfig.lib.json"
+        },
+        {
+            "path": "tsconfig.test.json"
+        }
+    ]
+}
