+++ packages/utils/tsconfig.test.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig.test.json",
+    "references": [
+        {
+            "path": "./tsconfig.lib.json"
+        }
+    ]
+}
