+++ packages/utils/vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+    test: {
+        name: "utils",
+        include: ["tests/**/*.test.{ts,tsx}"],
+    },
+});
