+++ tsconfig.json
@@ -15,6 +15,9 @@
             "path": "packages/mcp/tsconfig.json"
         },
         {
+            "path": "packages/utils/tsconfig.json"
+        },
+        {
             "path": "packages/native/tsconfig.json"
         },
         {
